+++ src/main/java/one/talon/api/model/ApplicationEvent.java
@@ -198,6 +198,15 @@ public class ApplicationEvent {
   **/
   @ApiModelProperty(required = true, value = "An array containing the effects that were applied as a result of this event.")
   public List<List<Object>> getEffects() {
+    for (int i = 0; i < effects.size(); i++) {
+      Double campaignid = Double.parseDouble(effects.get(i).get(0).toString());
+      Double rulesetid = Double.parseDouble(effects.get(i).get(1).toString());
+      Double ruleindex = Double.parseDouble(effects.get(i).get(2).toString());
+
+      effects.get(i).set(0, campaignid.intValue());
+      effects.get(i).set(1, rulesetid.intValue());
+      effects.get(i).set(2, ruleindex.intValue());
+    }
     return effects;
   }