+++ src/main/java/one/talon/api/model/Rule.java
@@ -147,6 +147,15 @@ public class Rule {
   **/
   @ApiModelProperty(required = true, value = "An array of effectful Talang expressions in arrays that will be evaluated when a rule matches.")
   public List<List<Object>> getEffects() {
+    for (int i = 0; i < effects.size(); i++) {
+      Double campaignid = Double.parseDouble(effects.get(i).get(0).toString());
+      Double rulesetid = Double.parseDouble(effects.get(i).get(1).toString());
+      Double ruleindex = Double.parseDouble(effects.get(i).get(2).toString());
+
+      effects.get(i).set(0, campaignid.intValue());
+      effects.get(i).set(1, rulesetid.intValue());
+      effects.get(i).set(2, ruleindex.intValue());
+    }
     return effects;
   }